+++ GustyLib/GustyLib/Core/classes/3rd party/IBActionSheet/IFA_IBActionSheet.m
@@ -140,7 +140,7 @@ otherButtonTitlesArray:(NSArray *)otherTitlesArray iPadWidth:(CGFloat)iPadWidth 
             [topButton setTitle:[titles objectAtIndex:0] forState:UIControlStateAll];
             [self.buttons insertObject:topButton atIndex:0];
             
-            int whereToStop = titles.count - 1;
+            NSUInteger whereToStop = titles.count - 1;
             for (int i = 1; i < whereToStop; ++i) {
                 IBActionSheetButton *middleButton = [[IBActionSheetButton alloc] initWithMaximumActionSheetWidth:[self maximumActionSheetWidth]];
                 [middleButton setTitle:[titles objectAtIndex:i] forState:UIControlStateAll];
@@ -214,7 +214,7 @@ otherButtonTitlesArray:(NSArray *)otherTitlesArray iPadWidth:(CGFloat)iPadWidth 
         
         CGPoint pointOfReference = CGPointMake(CGRectGetWidth(self.frame) / 2.0, CGRectGetHeight(self.frame) - 30);
         
-        int whereToStop;
+        NSInteger whereToStop;
         if (self.hasCancelButton) {
             [self addSubview:[self.buttons lastObject]];
             [[self.buttons lastObject] setCenter:pointOfReference];
@@ -227,7 +227,7 @@ otherButtonTitlesArray:(NSArray *)otherTitlesArray iPadWidth:(CGFloat)iPadWidth 
             whereToStop = self.buttons.count - 1;
         }
         
-        for (int i = 0, j = whereToStop; i <= whereToStop; ++i, --j) {
+        for (NSInteger i = 0, j = whereToStop; i <= whereToStop; ++i, --j) {
             [self addSubview:[self.buttons objectAtIndex:i]];
             [[self.buttons objectAtIndex:i] setCenter:CGPointMake(pointOfReference.x, pointOfReference.y - (44.5 * j))];
         }
@@ -261,7 +261,7 @@ otherButtonTitlesArray:(NSArray *)otherTitlesArray iPadWidth:(CGFloat)iPadWidth 
 
         CGPoint pointOfReference = CGPointMake(CGRectGetWidth(self.frame) / 2.0, CGRectGetHeight(self.frame) - 30);
 
-        int whereToStop;
+        NSInteger whereToStop;
         if (self.hasCancelButton) {
             [self addSubview:[self.buttons lastObject]];
             [[self.buttons lastObject] setCenter:pointOfReference];
@@ -274,7 +274,7 @@ otherButtonTitlesArray:(NSArray *)otherTitlesArray iPadWidth:(CGFloat)iPadWidth 
             whereToStop = self.buttons.count - 1;
         }
 
-        for (int i = 0, j = whereToStop; i <= whereToStop; ++i, --j) {
+        for (NSInteger i = 0, j = whereToStop; i <= whereToStop; ++i, --j) {
             [self addSubview:[self.buttons objectAtIndex:i]];
             [[self.buttons objectAtIndex:i] setCenter:CGPointMake(pointOfReference.x, pointOfReference.y - (45 * j))];
         }
@@ -346,7 +346,7 @@ otherButtonTitlesArray:(NSArray *)otherTitlesArray iPadWidth:(CGFloat)iPadWidth 
 
 - (NSInteger)addButtonWithTitle:(NSString *)title {
     
-    int index = self.buttons.count;
+    NSInteger index = self.buttons.count;
     
     if (self.hasCancelButton) {
         index -= 1;
